Fixed issue where it was possible to get the wrong caption
