+++ src/localization/captions/spec/CaptionTestData.js
@@ -1,0 +1,47 @@
+export default {
+  HelloWorld: [
+    {
+      content: 'hello',
+      start: 0,
+      end: 1200
+    },
+    {
+      content: 'world',
+      start: 1300,
+      end: 2500
+    }
+  ],
+  Other: [
+    {
+      content: 'things',
+      start: 0,
+      end: 1200
+    }
+  ],
+  Html: [
+    {
+      content: '<b>hello</b>',
+      start: 0,
+      end: 1200
+    }
+  ],
+  Template: [
+    {
+      content: '<b>{{greeting}}</b>',
+      start: 0,
+      end: 1200
+    }
+  ],
+  SameTime: [
+    {
+      content: 'foo',
+      start: 0,
+      end: 1000
+    },
+    {
+      content: 'bar',
+      start: 1000,
+      end: 1000
+    }
+  ]
+};
